requirements.txt
Django==2.2.15
Pillow==7.2.0
pytz==2020.1
sorl-thumbnail==12.4.1
sqlparse==0.3.1
